requirements.txt
scipy
librosa
essentia-tensorflow
numpy
pytest
flask
python-dotenv